chore: add utils directory
