+++ cypress/component/List.spec.tsx
@@ -3,15 +3,7 @@ import { List, Referrals16 } from '@toptal/picasso'
 import { mount } from '@cypress/react'
 import { TestingPicasso } from '@toptal/picasso/test-utils'
 
-/* eslint-disable react/no-array-index-key */
-const generateListItems = (total: number, listItemProps?: any) =>
-  Array(total)
-    .fill(0)
-    .map((_, index) => (
-      <List.Item key={index} {...listItemProps}>
-        {`list item N${index + 1}`}
-      </List.Item>
-    ))
+import { generateListItems } from '../../packages/picasso/src/List/utils'
 
 describe('List', () => {
   describe('Unordered', () => {
