+++ packages/picasso/src/List/utils/index.ts
@@ -1,0 +1,3 @@
+import generateListItems from './generateListItems'
+
+export { generateListItems }
